Exceptions Handled

+++ src/main/java/com/tech/java/springvihangaRest/Services/DepartmentServiceImpl.java
@@ -11,14 +11,14 @@ import java.util.Objects;
 import java.util.Optional;
 
 @Service
-public class DepartmentServiceImpl implements DepartmentService{
+public class DepartmentServiceImpl implements DepartmentService {
 
     @Autowired
     private DepartmentRepository departmentRepository;
 
     @Override
     public Department saveDepartment(Department department) {
-       return departmentRepository.save(department);
+        return departmentRepository.save(department);
     }
 
     @Override
@@ -27,38 +27,55 @@ public class DepartmentServiceImpl implements DepartmentService{
     }
 
     @Override
-    public Department updateDepartment(Long departmentId,Department department) {
-        Department DepartmentfrmDB= departmentRepository.findById(departmentId).get();
+    public Department updateDepartment(Long departmentId, Department department) throws DepartmentNotFoundException {
+        Optional<Department> DepartmentfrmDB = departmentRepository.findById(departmentId);
 
-        if(Objects.nonNull(department.getDepartmentName()) && !"".equalsIgnoreCase(department.getDepartmentName())){
-            DepartmentfrmDB.setDepartmentName(department.getDepartmentName());
+        if (!DepartmentfrmDB.isPresent()) {
+            throw new DepartmentNotFoundException("Department with the ID no: " + departmentId + " not found");
+        } else {
+
+            if (Objects.nonNull(department.getDepartmentName()) && !"".equalsIgnoreCase(department.getDepartmentName())) {
+                DepartmentfrmDB.get().setDepartmentName(department.getDepartmentName());
+            }
+
+            if (Objects.nonNull(department.getDepartmentAddress()) && !"".equalsIgnoreCase(department.getDepartmentAddress())) {
+                DepartmentfrmDB.get().setDepartmentAddress(department.getDepartmentAddress());
+            }
+
+            if (Objects.nonNull(department.getDepartmentCode()) && !"".equalsIgnoreCase(department.getDepartmentCode())) {
+                DepartmentfrmDB.get().setDepartmentCode(department.getDepartmentCode());
+            }
         }
-
-        if( Objects.nonNull(department.getDepartmentAddress()) && !"".equalsIgnoreCase(department.getDepartmentAddress())){
-            DepartmentfrmDB.setDepartmentAddress(department.getDepartmentAddress());
-        }
-
-        if(Objects.nonNull(department.getDepartmentCode()) && !"".equalsIgnoreCase(department.getDepartmentCode())){
-            DepartmentfrmDB.setDepartmentCode(department.getDepartmentCode());
-        }
-
-        return departmentRepository.save(DepartmentfrmDB);
+        return departmentRepository.save(DepartmentfrmDB.get());
 
     }
 
     @Override
-    public Department getDepamentByID(Long departmentId) {
-        return departmentRepository.findById(departmentId).get();
+    public Department getDepamentByID(Long departmentId) throws DepartmentNotFoundException {
+        Optional<Department> department = departmentRepository.findById(departmentId);
+        if (!department.isPresent()) {
+            throw new DepartmentNotFoundException("Department with the ID:" + departmentId + " is Not Available");
+        }
+        return department.get();
     }
 
     @Override
-    public void deleteDepartmentById(Long departmentId) {
-        departmentRepository.deleteById(departmentId);
+    public void deleteDepartmentById(Long departmentId) throws DepartmentNotFoundException {
+        Optional<Department> DBdepartment = departmentRepository.findById(departmentId);
+        if (!DBdepartment.isPresent()) {
+            throw new DepartmentNotFoundException("Department with the ID:" + departmentId + " is Not Available");
+        } else {
+            departmentRepository.deleteById(departmentId);
+        }
     }
 
     @Override
-    public Department getDepartmentByName(String departmentName) {
-        return departmentRepository.findByDepartmentNameIgnoreCase(departmentName);
+    public Department getDepartmentByName(String departmentName) throws DepartmentNotFoundException {
+        Optional<Department> DepartmentFrmDB = departmentRepository.findByDepartmentNameIgnoreCase(departmentName);
+        if (!DepartmentFrmDB.isPresent()) {
+            throw new DepartmentNotFoundException("Department with name " + departmentName + " is not found");
+        }
+        return DepartmentFrmDB.get();
     }
